refactor: move changed-file filtering into engine
